requirements.txt
streamlit==1.32.0
Pillow==10.1.0
requests==2.31.0
piexif==1.1.3
python-dotenv==1.0.0